Also look for windows path slash in java source name.
We guess some of the parent directories based on object's package name.
If the java source file contains any slashes, disable this heuristic.

As far as I know java compiler currently never includes any directories,
so this is just forward-looking change to make sure we do not generate
odd paths if the java compiler ever changes.

Change-Id: I7db7b4409e8eac0cbbabad00abf42e1c977e9669