WIP: Add infrastructure for merging Slow Paths.
This patch adds infrastructure for merging SlowPaths. Currently this
is still experimental and for now only applied to the ARM64 backend.

A list of the current categories of SlowPaths:
* Fatal SlowPaths - SlowPaths that do not return to the generated code
  - e.g DivZeroCheckSlowPath.
  - can be shared as long as the runtime can identify the SlowPath caller.
* Non-Fatal SlowPaths - that do not contain any parallel moves
  - e.g SuspendCheckSlowPath
  - can be shared as long as the runtime can identify the SlowPath caller
  and the shared SlowPath saves the union of live registers from each
  call path.
* Non-Fatal SlowPaths - that contain parallel moves.
  - e.g LoadStringSlowPath
  - can be shared as long as the runtime can identify the SlowPath caller,
  the shared SlowPath saves the union of live registers from each call
  path and the paralell moves

The framework introduced in this patch only tries to share SlowPaths
from the first two categories. However, once finished the framework
should be easily applicable to the third category (probably by running a
post register allocation pass that tries to share any Non-Fatal
SlowPaths that contain the same parallel moves).

Here is a list of the infrastructure changes introduced by this patch:
* The SlowPathCode class is a first citizen now, containing a list of
users (HInstructions) as well as a RegisterSet.
* The SlowPath creation has been moved to the LocationsBuilder and every
HInstruction now has a SlowPath member. This way we can now take the
decision of sharing the SlowPaths and determining the correct frame
size before the CodeGenerator visitor (the frame size depends on the
union of live registers from each SlowPath caller).
* For Non-Fatal SlowPaths that have been shared we call into the
SlowPath using a BL (branch and link) and return with a RET. Besides
saving the live registers (the union of the live registers from each
call path) we also save the LR (link register) at a known stack
location (first slot after the callee saves). This extra information
(saved link register) will help the runtime code retrieve the correct
StackMap (the one associated with the correct SlowPath caller).
* StackMaps now contain an extra filed - native_lr_offset. When
available this will be associated with the PC of the SlowPath caller.
* A clean-up of the current ARM64 SlowPaths. They do not rely anymore on
having been passed a Location (that would otherwise be invalid at
LocationsBuilder time).

What is missing:
* Runtime support for dealing with non-fatal SlowPaths (StackMaps that
contain a native_lr filed). In short where we previously were looking
for a matching native_pc_offset, now, if the StackMap has a
native_lr_offset there might be duplicate StackMaps for the current
native_pc_offset. In this case we should load the current native_lr
form the known stack location and find the matching StackMap (that
matches both the native_pc and the native_lr offsets). Since there is
still room for improvement on the StackMaps this has not been implemented
in this patch.
* Checker tests infrastructure for asserting that the slowpath sharing
has/has not happened (based on SlowPathCode->GetDescription()).
* Support for sharing non-fatal SlowPaths that contain parallel moves.

What needs improvement:
* There are still improvements to be done on the StackMaps compression.
Currently even though this patch reduces the .text section of the oat
files, there is a considerable increase in the .rodata section (where
the StackMaps live). Most of the wasted space is associated with
StackMaps that do not need a native_lr_offset, but are part of a method
that shares at least a non-fatal SlowPath (thus requiring
number_of_bytes_for_native_lr for each SlowPath within that method).

Size measurements on boot.oat:
  * Before (without SlowPaths sharing):
    section             size
    .rodata             39903232
    .text               36918852

  * After (with SlowPaths sharing):
    section             size
    .rodata             40275968
    .text               36886068

Please note that this patch will be split into a smaller, more digestible
series, once the whole infrastructure is in place and the above numbers
look better for both elf sections.

Change-Id: If4981c53c5b9760f7942a66465a52c0a02598cbf
Signed-off-by: Serban Constantinescu <serban.constantinescu@linaro.org>
Signed-off-by: Ramin Zaghi <ramin.zaghi@arm.com>