WIP: Add infrastructure for merging Slow Paths.
This patch adds infrastructure for merging "fatal" slowpaths. These are
slowpaths that do not return to the generated code. The framework
introduced here takes the decision of sharing slowpaths if two fatal
slowpaths of the same type exist in the same compilation unit.

This is the base for a more complicated patch that would be able to share
slowpaths that return back to the generated code (e.g. SuspendChecks).

Currently we decide if we need or not a slowpath at codegen time. This
however impeeds us from allocating the correct stack size if we decide
to share slowpaths with a different live register set (thus the shared
slowpath will be saving the union of live registers). For this reason
we have moved the slowpath creation for DivZeroCheck and NullCheck to
the locations builder visitors (though we do not yet use this
infrastructure the reviewer should be aware of the incoming changes).

Once the disassembly infrastructure has been merged, we should extend
the checker tests and apply them to the slowpaths sharing tests.

Change-Id: If4981c53c5b9760f7942a66465a52c0a02598cbf
Signed-off-by: Serban Constantinescu <serban.constantinescu@linaro.org>
Signed-off-by: Ramin Zaghi <ramin.zaghi@arm.com>