Set default visibility on art_sigsegv_fault.
libart.so is built with flag -fvisibility=protected
which complicates setting breakpoint on art_sigsegv_fault.
Accordingly to GCC manual "Protected visibility indicates
that the symbol will be placed in the dynamic symbol table,
but that references within the defining module will bind
to the local symbol."
So, breakpoint on protected symbol will not be hit.
Without this change gdbclient with enabled ART debugging
doesn't stop art_sigsegv_fault when app crashes with SIGSEGV.

Change-Id: Ib73344e710966eddf55a6facb29558504a511f00