Do not log mmaps if dlopen fails due to known reason.
If LD_USE_LOAD_BIAS is not set, dlopen will not load the ELF file
at the fixed address and we will fallback to our custom loader.
Since this is know path, do not try to use dlopen and do not log
the mmaps.  We will still try and log if dlopen fails due to
some reason that we are not aware of.

Change-Id: Ifbbf27a4e3987f22c149a1f297544aa7f5e3a8fc