Added index type of dex byte instructions.
Rationale:
The information was already in the instruction_list but not
exposed as enum through a getter. This CL adds that ability.
This information is needed by the upcoming Art-based dexdump.

NOTE 1:
I added all enums that were in the original Dalvik list,
even though some are not used yet.

NOTE 2:
I used the naming convention found the in list, although
I would have a slight preference for using Index in the
enum names (kIndexNone, kIndexTypeRef, etc.).

Bug: 17442393

Change-Id: Ica18ae43135f78f29e9832f5a101d497e54724cf