Optimizing String.Equals as an intrinsic (ARM)
The second implementation of String.Equals. I added an intrinsic
in ARM which is similar to the original java implementation
of String.equals: an instanceof check, null check, length check, and
reference equality check followed by a loop comparing strings character
by character. This should be faster than the original implementation
because I am removing the overhead of storing/loading caller and callee
saved registers which is necessary in the java implementation during calls
to CharAt(). Performance benchmarks coming soon.

Bug: 21481923
Change-Id: I48df2a74f2a92b56fb9479fbf14276d44e880aed