[DO NOT MERGE] Add JNI state change performance test
I just upload this test for my record. This test is only for measuring
the JNI performance manually. The test prints timing information on
each run so we cannot compare stdout with expected.txt.

Also made libarttest.so link with libart.so instead of libartd.so.
This affects all tests that add custom JNI functions in libarttest.so.

Change-Id: I68aa1473c425d9ef6f00796bdb47bea2a956d011