Support X86 intrinsic System.arraycopy char
This is an implementation of arraycopy for X86 and X86_64 versions
of intrinsic System.arraycopy(char[], int, char[], int, int).

The implementations use rep movsw to copy the chars.

This also fixes a bug in the register allocator, where Temps that had
explicit registers didn't get added to the AllocatedRegisters set, and
therefore were not preserved in the prologue/epilogue.

Change-Id: Icf9d0efb9986bc3e0794238a74f94fe02f9b42be
Signed-off-by: Mark Mendell <mark.p.mendell@intel.com>