Add a GVN dependency 'GC' for garbage collection.
This will be used by incoming architecture specific optimisations.

The dependencies must be conservative. When an HInstruction is
created we may not be sure whether it can trigger GC. In that
case the 'ChangesGC' dependency must be set. We control at
code-generation time that any HInstruction that can call has the
'ChangesGC' dependency set.

Change-Id: Iea6a7f430009f37a9599b0a0039207049906e45d