[WIP] ART: Add more support for Clang integrated as
The Clang assembler still does not support .altmacro. However,
the Clang preprocessor works different wrt/ the GCC preprocessor
and does not give us trouble with inserted spaces.

Add paths in the asm support for x86 and x86-64 to that allow
to use the integrated assembler.

Remove conditions for old, no longer supported versions of Clang.

Hope would be that at some point we can remove all the index
nonsense, that is, the horrible dual-casing now only necessary
because of the Mac.

Bug: 17443165
Change-Id: I1a077d4b612abc2b1b851c1bdabb5008a52e5aa6