Enable ASan annotations in ART based on macro setting.
This is a quick hack to remove ASan annotations from SANITIZE_TARGET
build, because they are breaking everything - it looks like we forget
to clean redzones from a page before releasing it to the system, and
some unrelated code runs into them later.

We will add some way to control this from the build system later,
or just fix the underlying issue and re-enable by default.

Change-Id: If79e2455e246dc39355f2758323b249ac20b6144