Fix Quick/Optimizing suspend check assumption mismatch.
Quick expects that every method contains a suspend check, so
it eliminates suspend checks in loops containing an invoke.
However, Optimizing eliminates the suspend check from leaf
methods, so the combination of a Quick-compiled loop calling
an Optimizing-compiled leaf method can lead to missing
suspend checks and potentially leading to ANRs.

Currently, all methods containing a try-catch are compiled
with Quick, so it's relatively easy to create a regression
test. However, this test will not be valid when Optimizing
starts supporting try-catch.

TODO: FIX! So far, only the test is provided.

Bug: 22657404
Change-Id: I3bc40bf3f5c1e7d18704d1547b139e939950b770