[WIP] ART: Remove Thread::Current from compiler
Profiling shows excessive usage of Thread::Current in the compiler,
which on device can lead to overhead (5% on arm32). Instead, supply
the thread from the compiler driver through the compiler stages.

This reduces the Thread::Current calls from >10M to <1M for
compiling an arm32 boot image. Most of it, >500K, is taken up by
arena allocators.

TODO: Fix the tests.
TODO: Consider making the thread always the first argument to be
      consistent with other places.

Change-Id: Ic3f26db4e352cc7924871fccfae83ffd55a2cf01