[WIP] ART: Remove Thread::Current from compiler
Profiling shows excessive usage of Thread::Current in the compiler,
which on device can lead to overhead (5% on arm32). Instead, supply
the thread from the compiler driver through the compiler stages.

This reduces the Thread::Current calls from >10M to ~6.7M for
compiling an arm32 boot image.

TODO: Fix the tests.
TODO: Consider making the thread always the first argument to be
      consistent with other places.
TODO: Look into where the rest of the calls are made. Probably this is
      the class-linker. Quick try for investigation didn't work so
      well, DumpNativeStack died. Consider backtrace.h.

Change-Id: Ic3f26db4e352cc7924871fccfae83ffd55a2cf01