WIP: Add read barrier instrumentation in Quick
This is a purely experimental attempt to add read barrier in Quick,
as we are unlikely to support Quick with read barrier in the future.
Adding read barrier to Quick might help testing the read barrier.

Currently only added partial implementation for X86/64, not fully tested
and not sure if it works correctly.

Change-Id: I7f46049be3f57fa882bfb6510eff2266b8e8308a