ARM64: Instruction simplification for array accesses.
HArrayGet and HArraySet with variable indexes generate two
instructions on arm64, like

    add temp, obj, #data_offset
    ldr out, [temp, index LSL #shift_amount]

When we have multiple accesses to the same array, the initial `add`
instruction is redundant.

This patch introduces an architecture-specific instruction
simplification pass. Eventually, different backends can implement
different transformations on the graph there.
For now, on arm64 it splits HArrayGet and HArraySet using the new
arm64-specific IR HIntermediateAddress. After that we run GVN again to
squash the multiple occurrences of HIntermediateAddress.

Change-Id: I2e3d12fbb07fed07b2cb2f3f47f99f5a032f8312