Reduce lock contention on JIT instrumentation lock
Steal 16 bits from ArtMethod::method_index_ field to make an ID field
for ArtMethod. The ID is initialized by JIT only for ArtMethod instances
that have executed at least once. This doesn't add space overhead.

JIT uses this ID as an index to a counter array, to avoid the use of
a global lock. If the ID exceeds 64K, JIT falls back to use the old
hash map with the global lock. The counter array adds a fixed overhead
of 64K*sizeof(size_t).

Change-Id: Iee9acf86f859a37a2b451df55b0e84657ec89faa