Svelter libart-compiler
Added new environment variable ART_{TARGET,HOST}_CODEGEN_ARCHS which may be set to 'all', 'svelte'
or a space separated list of architectures.

When compiled with ART_{TARGET,HOST}_CODEGEN_ARCHS='all' (the default value) dex2oat will be able
to generate output for all supported architectures.

When compiled with ART_TARGET_CODEGEN_ARCHS='svelte' only the architectures of the TARGET will be
included. When ART_HOST_CODEGEN_ARCHS='svelte' all architectures the target includes and the host
architectures will be included on the host dex2oat.

If a list of architectures is given only those will be included.

Change-Id: I87f4ad0131ab1b37544d8799e947ce4733b6daec