Don't allow GC to start if the runtime is shutting down
Daemon threads may be trying to do allocations as the runtime is
shutting down, don't allow GC to start since it touches the heap
and may cause crashes if it is running while we delete the heap
or thread list.

TODO: Is it OK to possibly throw OOM error here?

Bug: 18577101
Change-Id: I1404b41f2aee13d3c26acffdf588c985628beefc