[WIP] Add read barrier in optimizing compiler for X86
This is sort of experimental since we haven't decided which is
better for the long term: to add read barrier in the HIR, or in each
back-end's code generator. This CL only inserts the read barrier slow path
calls, but not inlining the fast path in compiler.

Currently I've added read barrier for all places that have heap unpoisoning,
except for GenCAS() for intrinsics. There are some places that needs a
different read barrier for root, such as in VisitLoadClass(). I haven't
added these yet.

I use a kForceReadBarrier flag to force inserting the read barrier calls
in optimizing compiler, but I still use the CMS collector. This enables us
to test if the calls are inserted correctly, without using the CC collector.
The CC collector would likely to fail as we don't have read barriers for the
roots (mentioned above).

Change-Id: Id7179b1ef338ecdd4f82071d32c250eb6a9cabeb