Fix deoptimization with pending exception
When deoptimizing the stack, we set a fake exception in the current
Thread* (see method Thread::GetDeoptimizationException). On the next
exception check, the QuickExceptionHandler will deoptimize the stack.

The issue is when we deoptimize while an exception is already pending
in the current Thread*: setting the fake exception will clobber the
pending exception which is not correct. This happens in the
artQuickToInterpreterBridge when returning from the interpreter and
we want to deoptimize the stack for debugging (like single-stepping).

This CL saves the pending exception before asking for deoptimization.
Then the exception is restored just before executing the deoptimized
frames with the interpreter.

This CL also contains the following cleanup:
- Renames class DeoptimizationReturnValueRecord to class
  DeoptimizationContextRecord since it now handles the return value
  the exception.
- Remove unused method Thread::IsDeoptimizationReturnValueReference.

TODO: check we do visit objects (preserved for deopt) for GC

Bug: 23371176
Change-Id: I7f4c8347b328817c452beda3399e210eba3a88a4